requirements.txt
Flask==3.1.2
google-genai==0.2.2
python-dotenv==1.0.0